% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winr-package.R
\docType{package}
\name{winr-package}
\alias{winr}
\alias{winr-package}
\title{winr: Randomization Based Methods for Covariance and Stratified Adjustment of Win Ratios and Win Odds for Ordinal Outcomes}
\description{
A multi-visit clinical trial may collect participant responses to interventions on an ordinal, rather than interval, measurement scale. The study may also have a stratified design, such as randomization within centers and collect information at multiple visits to evaluate treatment efficacy. Baseline characteristics may be strongly associated with the outcome, and adjustment for them can improve power. The win ratio (ignores ties) and the win odds (accounts for ties) can be useful when analyzing these types of data from randomized controlled trials. This package provides straightforward methods for stratification and baseline covariate adjusted win ratio and win odds estimates for a multi-visit clinical trial to compare a test treatment to a control treatment.
}
\author{
\strong{Maintainer}: Ann Marie Weideman \email{anndo1@umbc.edu} (\href{https://orcid.org/0000-0002-1653-9102}{ORCID})

Authors:
\itemize{
  \item Elaine Kowalewski \email{elainekkowalewski@gmail.com} (\href{https://orcid.org/0000-0002-7915-877X}{ORCID})
  \item Gary Koch \email{BCL@bios.unc.edu}
}

}
\keyword{internal}
